sync

+++ naming/src/main/java/com/alibaba/nacos/naming/core/v2/metadata/NamingMetadataManager.java
@@ -69,7 +69,7 @@ public class NamingMetadataManager extends SmartSubscriber {
     }
     
     /**
-     * Whether instance metadata metadata for instance of {@link Service}.
+     * Whether instance metadata for instance of {@link Service}.
      *
      * @param service    service
      * @param metadataId instance metadata id
@@ -82,7 +82,7 @@ public class NamingMetadataManager extends SmartSubscriber {
     /**
      * Get service metadata for {@link Service}, which is the original metadata object.
      *
-     * <p>This method should use only query, can't modified metadata.
+     * <p>This method should use only query, can't modify metadata.
      *
      * @param service service
      * @return service metadata
@@ -94,7 +94,7 @@ public class NamingMetadataManager extends SmartSubscriber {
     /**
      * Get instance metadata for instance of {@link Service}, which is the original metadata object.
      *
-     * <p>This method should use only query, can't modified metadata.
+     * <p>This method should use only query, can't modify metadata.
      *
      * @param service    service
      * @param metadataId instance metadata id
@@ -151,9 +151,11 @@ public class NamingMetadataManager extends SmartSubscriber {
      */
     public void removeInstanceMetadata(Service service, String metadataId) {
         ConcurrentMap<String, InstanceMetadata> instanceMetadataMapForService = instanceMetadataMap.get(service);
-        instanceMetadataMapForService.remove(metadataId);
-        if (instanceMetadataMapForService.isEmpty()) {
-            serviceMetadataMap.remove(service);
+        if (null != instanceMetadataMapForService) {
+            instanceMetadataMapForService.remove(metadataId);
+            if (instanceMetadataMapForService.isEmpty()) {
+                serviceMetadataMap.remove(service);
+            }
         }
         expiredMetadataInfos.remove(ExpiredMetadataInfo.newExpiredInstanceMetadata(service, metadataId));
     }
